Extracts questions handling code into its own component
